requirements-test.txt
molecule
molecule-plugins[docker]
docker
yamllint
ansible-lint
andebox
pytest-ansible
